+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## (TBA)
 
 - Enhancements
-    - Added possibility to use rich Text objects to set the description argument of a `Settanble`
+    - Added possibility to use rich Text objects to set the description argument of a `Settable`
 
 ## 4.1.1 (July 9, 2026)